Use Criterion, not libtest, for benchmarking
This patch replaces the usage of the crate `test` ("libtest") for
benchmarking, which requires the nightly-only `test` feature, with the
crate `criterion` ("Criterion"), as suggested by @jedisct1, which
works on stable Rust.

As the `test` feature was the only nightly Rust feature used by
`clockpro_cache`, this patch enables `clockpro_cache` to work on
stable Rust, as it had at version 0.1.6.

I have tested that the Criterion-based benchmark introduced by this
patch does appear to work; however, I am unfamiliar with nightly Rust
and would not be able to tell whether any significant differences
exist between the benchmarking provided by Criterion and that provided
by libtest.

GitHub: Resolve jedisct1/rust-clockpro-cache#5